% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Server function}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Input}

\item{output}{Output. For details, checkout Shiny documentation}

\item{session}{Session. For details, checkout Shiny documentation}
}
\description{
Server function
}
