% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_bivariate_plot.R
\name{bivariate_server}
\alias{bivariate_server}
\title{either plots a whole bivariate plot or a bivariate plot animation through time.
param @id: to link the input and the output
param @variables: the variables to plot
param @data: the data to work with}
\usage{
bivariate_server(id, variables, data)
}
\description{
either plots a whole bivariate plot or a bivariate plot animation through time.
param @id: to link the input and the output
param @variables: the variables to plot
param @data: the data to work with
}
