% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_bivariate_plot.R
\name{bivariate_ui}
\alias{bivariate_ui}
\title{lets the user select the variables to be graphed
param @id: to link input and output
param @xvar: the x variables
param @yvar: the y variables}
\usage{
bivariate_ui(id, xvar, yvar)
}
\description{
lets the user select the variables to be graphed
param @id: to link input and output
param @xvar: the x variables
param @yvar: the y variables
}
