% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_preview.R
\name{data_preview_server}
\alias{data_preview_server}
\title{Constructs the path to the file chosen}
\usage{
data_preview_server(id)
}
\arguments{
\item{id}{to link the input and the output}

\item{start_path}{starting path to be added to the full path}
}
\value{
the loaded data
}
\description{
Constructs the path to the file chosen
}
