% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_preview.R
\name{data_preview_var_ui}
\alias{data_preview_var_ui}
\title{lets the user select the site, season and date of the file}
\usage{
data_preview_var_ui(id, datasets, site)
}
\arguments{
\item{id}{a string to link input and output}

\item{datasets}{a named list of the columns of the data}

\item{site}{a named list of possible sites that the user can choose}
}
\value{
a taglist of input
}
\description{
lets the user select the site, season and date of the file
}
