% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fit_eval}
\alias{fit_eval}
\title{Evaluating lm fit on full data and evaluating}
\usage{
fit_eval(data, formula)
}
\arguments{
\item{data}{The data as data frame}

\item{formula}{Regression formula class}
}
\value{
A list of objects including model summary, fitting plots and residuals
}
\description{
Evaluating lm fit on full data and evaluating
}
