% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_data_preview.R
\name{full_data_preview_server}
\alias{full_data_preview_server}
\title{Server function for pre-loading data files}
\usage{
full_data_preview_server(id)
}
\arguments{
\item{id}{to link the input and the output}
}
\value{
the loaded data
}
\description{
Server function for pre-loading data files
}
