% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_data_preview.R
\name{full_data_preview_ui}
\alias{full_data_preview_ui}
\title{lets the user select the site, season and date of the file}
\usage{
full_data_preview_ui(id)
}
\arguments{
\item{id}{a string to link input and output}
}
\value{
a taglist of input
}
\description{
lets the user select the site, season and date of the file
}
