% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_var_vs_time.R
\name{full_time_plot_server}
\alias{full_time_plot_server}
\title{returns a ggplot graph depending on which type of graph was selected}
\usage{
full_time_plot_server(id, data, varlist)
}
\arguments{
\item{id:}{id that connects the ui to the server}

\item{data:}{the data to work with. a reactive variable}

\item{variables:}{the variables to graph to make nice axis}
}
\description{
returns a ggplot graph depending on which type of graph was selected
}
