% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_var_vs_time.R
\name{full_time_plot_ui}
\alias{full_time_plot_ui}
\title{lets the user select the variables to be graphed
param @id: to link ui and server
param @y_variables: y variables to select
param @x_variables: x variables to select}
\usage{
full_time_plot_ui(id, y_variables, x_variables)
}
\description{
lets the user select the variables to be graphed
param @id: to link ui and server
param @y_variables: y variables to select
param @x_variables: x variables to select
}
