% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{full_var_times}
\alias{full_var_times}
\title{Variables for the full data section}
\format{
An object of class \code{character} of length 4.
}
\usage{
full_var_times
}
\description{
Variables for the full data section
}
\keyword{datasets}
