% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gap_filling}
\alias{gap_filling}
\title{This function performs gapfilling and plotting of gapfilling variables}
\usage{
gap_filling(data, formula)
}
\arguments{
\item{data}{The data of interest}

\item{formula}{The model formula chosen to gapfill}
}
\description{
This function performs gapfilling and plotting of gapfilling variables
}
