% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_var_vs_time.R
\name{generate_vvtplt}
\alias{generate_vvtplt}
\title{Function to generate the plot using a skeleton plot, data and input}
\usage{
generate_vvtplt(
  id,
  input,
  output,
  session,
  title = "Variable vs Time",
  data,
  varlist
)
}
\arguments{
\item{input}{The input wrapped inside a server function}

\item{data}{The data set}
}
\description{
Function to generate the plot using a skeleton plot, data and input
}
