% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_var_fullname}
\alias{get_var_fullname}
\title{Getting full name of variables}
\usage{
get_var_fullname(varname, varlist)
}
\arguments{
\item{varname}{A vector of shortened name of variables such as 'w'}

\item{varlist}{A list of variables and their original codes}
}
\value{
A vector of matched full names
}
\description{
Getting full name of variables
}
