% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_histogram.R
\name{histogram_server}
\alias{histogram_server}
\title{Plot the histogram for the file chosen}
\usage{
histogram_server(id, variables, data)
}
\arguments{
\item{id}{to link the input and the output}

\item{variables}{the variables to plot -- used to get nice labels on axes}

\item{data}{the data to work with -- usually a reactive object}
}
\value{
a plotly object
}
\description{
Plot the histogram for the file chosen
}
