% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_histogram.R
\name{histogram_ui}
\alias{histogram_ui}
\title{lets the user select the variables to be graphed}
\usage{
histogram_ui(id, variables)
}
\arguments{
\item{id:}{to link input and output}

\item{variables:}{a named list of variables to choose from}
}
\value{
a taglist of input function
}
\description{
lets the user select the variables to be graphed
}
