% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{oos_prediction}
\alias{oos_prediction}
\title{Out of sample prediction models using train/test split}
\usage{
oos_prediction(data, formula, split_prop = 3/4)
}
\arguments{
\item{data}{Data of data.frame format}

\item{formula}{Regression formula class}

\item{split_prop}{Train-test split. Default 3/4 of the data set will be training}
}
\description{
Out of sample prediction models using train/test split
}
