% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_raw_data_bivariate}
\alias{plot_raw_data_bivariate}
\title{helper function to plot a whole bivariate plot or a bivariate plot animation through time.
param @input: list of input that user has chosen (given to us by shiny).
param @output: list of available ouput action (given to us by shiny).
param @variables: the variables to plot.
param @data: the data to work with.
param @season: which season to graph}
\usage{
plot_raw_data_bivariate(input, output, variables, data, season)
}
\description{
helper function to plot a whole bivariate plot or a bivariate plot animation through time.
param @input: list of input that user has chosen (given to us by shiny).
param @output: list of available ouput action (given to us by shiny).
param @variables: the variables to plot.
param @data: the data to work with.
param @season: which season to graph
}
