% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_var_vs_time.R
\name{plot_vvt_server}
\alias{plot_vvt_server}
\title{Server function for variable-versus-time (vvt) plotting}
\usage{
plot_vvt_server(id, data, varlist)
}
\arguments{
\item{id}{Identifier to match ui and server}

\item{data}{The data required}

\item{varlist}{List of variables}
}
\description{
Server function for variable-versus-time (vvt) plotting
}
