% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_var_vs_time.R
\name{plot_vvt_ui}
\alias{plot_vvt_ui}
\title{lets the user select the variables to be graphed}
\usage{
plot_vvt_ui(id, variables)
}
\arguments{
\item{id:}{to link input and output}
}
\description{
lets the user select the variables to be graphed
}
