% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{raw_variables}
\alias{raw_variables}
\title{Variables for the raw variables section}
\format{
An object of class \code{list} of length 6.
}
\usage{
raw_variables
}
\description{
Variables for the raw variables section
}
\keyword{datasets}
