% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_regression.R
\name{regression_server}
\alias{regression_server}
\title{Server function for regression module}
\usage{
regression_server(id, data)
}
\arguments{
\item{id}{Identifier for module, see Shiny Modules documentation}

\item{data}{The data being loaded}
}
\description{
Server function for regression module
}
