% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_regression.R
\name{regression_ui}
\alias{regression_ui}
\title{Variable selection for regression}
\usage{
regression_ui(id, variables)
}
\arguments{
\item{id}{Character used to specify namespace \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing ui elements
}
\description{
Variable selection for regression
}
