% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_var_vs_time.R
\name{vvt_selection}
\alias{vvt_selection}
\title{function to validate vvt input}
\usage{
vvt_selection(input)
}
\arguments{
\item{input}{the input from the ui}
}
\value{
NULL if input is validated or an error message
}
\description{
function to validate vvt input
}
