% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_bivariate_plot.R
\name{bivariate_server}
\alias{bivariate_server}
\title{Either plots a whole bivariate plot or a bivariate plot animation through time.}
\usage{
bivariate_server(id, variables, data)
}
\arguments{
\item{id:}{to link the input and the output}

\item{variables:}{the variables to plot}

\item{data:}{the data to work with}
}
\description{
Either plots a whole bivariate plot or a bivariate plot animation through time.
}
