% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_bivariate_plot.R
\name{bivariate_ui}
\alias{bivariate_ui}
\title{Lets the user select the variables to be graphed}
\usage{
bivariate_ui(id, xvar, yvar)
}
\arguments{
\item{id:}{to link input and output}

\item{xvar:}{the x variables}

\item{yvar:}{the y variables}
}
\description{
Lets the user select the variables to be graphed
}
